+++ jsmn.h
@@ -39,11 +39,11 @@ enum jsmnerr {
  */
 typedef struct {
 	jsmntype_t type;
-	int start;
-	int end;
-	int size;
+	long start;
+	long end;
+	long size;
 #ifdef JSMN_PARENT_LINKS
-	int parent;
+	long parent;
 #endif
 } jsmntok_t;
 
@@ -52,9 +52,9 @@ typedef struct {
  * the string being parsed now and current position in that string
  */
 typedef struct {
-	unsigned int pos; /* offset in the JSON string */
-	unsigned int toknext; /* next token to allocate */
-	int toksuper; /* superior token node, e.g parent object or array */
+	unsigned long pos; /* offset in the JSON string */
+	unsigned long toknext; /* next token to allocate */
+	long toksuper; /* superior token node, e.g parent object or array */
 } jsmn_parser;
 
 /**
@@ -66,8 +66,8 @@ void jsmn_init(jsmn_parser *parser);
  * Run JSON parser. It parses a JSON data string into and array of tokens, each describing
  * a single JSON object.
  */
-int jsmn_parse(jsmn_parser *parser, const char *js, size_t len,
-		jsmntok_t *tokens, unsigned int num_tokens);
+long jsmn_parse(jsmn_parser *parser, const char *js, size_t len,
+		jsmntok_t *tokens, unsigned long num_tokens);
 
 #ifdef __cplusplus
 }
